🔧 chore: 디렉토리 게층 변경
- inventory 서비스내에 존재하는 여러 도메인 별로 디렉토리를 재구성
