--- services/inventory-service/src/main/java/com/example/inventory/controller/HealthCheckController.java
@@ -1,22 +1,0 @@
-package com.example.inventory.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDateTime;
-import java.util.Map;
-
-@RestController
-public class HealthCheckController {
-
-    @GetMapping("/inventory/ping")
-    public ResponseEntity<Map<String, Object>> ping() {
-        return ResponseEntity.ok(Map.of(
-                "service", "inventory-service",
-                "status", "UP",
-                "timestamp", LocalDateTime.now()
-        ));
-    }
-
-}
