--- services/inventory-service/src/main/java/com/example/inventory/entity/inventory/DailyInventory.java
@@ -1,32 +1,0 @@
-package com.example.inventory.entity.inventory;
-
-import com.example.inventory.entity.product.Product;
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class DailyInventory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Setter
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate date; // 날짜별 일자
-
-    private int stockAvailable; // 가용 재고
-
-}
