+++ services/inventory-service/src/main/java/com/example/inventory/inventory/entity/DailyInventory.java
@@ -1,0 +1,31 @@
+package com.example.inventory.inventory.entity;
+
+import com.example.inventory.product.entity.Product;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDate;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class DailyInventory {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Setter
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "product_id", nullable = false)
+    private Product product;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+    private LocalDate date; // 날짜별 일자
+
+    private int stockAvailable; // 가용 재고
+
+}
