+++ services/inventory-service/src/main/java/com/example/inventory/product/repository/ProductRepository.java
@@ -1,0 +1,10 @@
+package com.example.inventory.product.repository;
+
+import com.example.inventory.product.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+    Product findByName(String name);
+
+}
