--- services/inventory-service/src/main/java/com/example/inventory/repository/DailyInventoryRepository.java
@@ -1,15 +1,0 @@
-package com.example.inventory.repository;
-
-import com.example.inventory.entity.inventory.DailyInventory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Repository
-public interface DailyInventoryRepository extends JpaRepository<DailyInventory, Long> {
-
-    List<DailyInventory> findByProductIdInAndDateBetween(List<Long> productIds, LocalDate checkIn, LocalDate checkOut);
-
-}
