+++ services/inventory-service/src/main/java/com/example/inventory/store/entity/Category.java
@@ -1,0 +1,18 @@
+package com.example.inventory.store.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Category {
+
+    HOTEL("호텔"),
+    PENSION("펜션"),
+    POOL_VILLA("풀빌라"),
+    GLAMPING("글램핑")
+            ;
+
+    private final String displayName;
+
+}
