+++ services/inventory-service/src/main/java/com/example/inventory/store/repository/StoreQueryDslRepository.java
@@ -1,0 +1,12 @@
+package com.example.inventory.store.repository;
+
+import com.example.inventory.store.dto.request.StoreUserFindRequest;
+import com.example.inventory.store.entity.Store;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface StoreQueryDslRepository {
+
+    Page<Store> findAllByFetchJoin(StoreUserFindRequest request, Pageable pageable);
+
+}
