+++ services/inventory-service/src/main/java/com/example/inventory/store/repository/StoreQueryDslRepositoryImpl.java
@@ -1,0 +1,148 @@
+    package com.example.inventory.store.repository;
+
+    import com.example.inventory.product.entity.Product;
+    import com.example.inventory.store.dto.request.StoreUserFindRequest;
+    import com.example.inventory.store.entity.Amenities;
+    import com.example.inventory.store.entity.Category;
+    import com.example.inventory.store.entity.Store;
+    import com.querydsl.core.types.dsl.BooleanExpression;
+    import com.querydsl.jpa.impl.JPAQueryFactory;
+    import lombok.RequiredArgsConstructor;
+    import org.springframework.data.domain.Page;
+    import org.springframework.data.domain.PageImpl;
+    import org.springframework.data.domain.Pageable;
+    import org.springframework.stereotype.Repository;
+    import org.springframework.transaction.annotation.Transactional;
+
+    import java.util.List;
+    import java.util.Map;
+    import java.util.Optional;
+    import java.util.stream.Collectors;
+
+    import static com.example.inventory.product.entity.QProduct.product;
+    import static com.example.inventory.store.entity.QStore.store;
+    import static com.example.inventory.util.QueryDslUtils.combineAnd;
+
+    @Repository
+    @RequiredArgsConstructor
+    public class StoreQueryDslRepositoryImpl implements StoreQueryDslRepository {
+
+        private final JPAQueryFactory jpaQueryFactory;
+
+        @Transactional(readOnly = true)
+        @Override
+        public Page<Store> findAllByFetchJoin(StoreUserFindRequest request, Pageable pageable) {
+            BooleanExpression whereCondition = getBooleanExpression(request);
+
+            List<Store> stores = jpaQueryFactory
+                    .select(store)
+                    .from(store)
+                    .leftJoin(store.amenities).fetchJoin()
+                    .where(whereCondition)
+                    .offset(pageable.getOffset())
+                    .limit(pageable.getPageSize())
+                    .distinct()
+                    .fetch();
+
+            List<Product> products = getProducts(stores);
+
+            mapProductsToStores(stores, products);
+
+            Long total = jpaQueryFactory
+                    .select(store.countDistinct()) // DISTINCT를 포함한 COUNT
+                    .from(store)
+                    .where(whereCondition)
+                    .fetchOne();
+
+            return new PageImpl<>(stores, pageable, Optional.ofNullable(total).orElse(0L));
+        }
+
+        private BooleanExpression getBooleanExpression(StoreUserFindRequest request) {
+            BooleanExpression nameCondition = nameLike(request.getName());
+            BooleanExpression categoryCondition = categoryEqual(request.getCategory());
+            BooleanExpression addressCondition = addressEqual(request.getProvince(), request.getCity());
+            BooleanExpression amenitiesCondition = amenitiesAllMatch(request.getAmenities());
+
+            return combineAnd(
+                    nameCondition,
+                    categoryCondition,
+                    addressCondition,
+                    amenitiesCondition
+            );
+        }
+
+        private List<Product> getProducts(List<Store> stores) {
+            List<Long> storeIds = stores.stream()
+                    .map(Store::getId)
+                    .toList();
+
+            return jpaQueryFactory.select(product)
+                    .from(product)
+                    .where(product.store.id.in(storeIds))
+                    .fetch();
+        }
+
+        private void mapProductsToStores(List<Store> stores, List<Product> products) {
+            Map<Long, List<Product>> productMap = products.stream()
+                    .collect(Collectors.groupingBy(p -> p.getStore().getId()));
+
+            stores.forEach(store -> {
+                List<Product> productList = productMap.get(store.getId());
+                store.assignProducts(productList);
+            });
+        }
+
+        private BooleanExpression nameLike(String name) {
+            if (name == null || name.isEmpty()) {
+                return null;
+            }
+
+            String searchName = "%" + name + "%";
+            return store.name.like(searchName);
+        }
+
+        private BooleanExpression categoryEqual(Category category) {
+            if (category == null) {
+                return null;
+            }
+
+            return store.category.eq(category);
+        }
+
+        private BooleanExpression addressEqual(String province, String city) {
+            if (province == null || province.isEmpty()) {
+                return null;
+            }
+
+            if(city == null || city.isEmpty()) {
+                return null;
+            }
+
+            return store.address.province.eq(province)
+                    .and(store.address.city.eq(city));
+        }
+
+        private BooleanExpression amenitiesAllMatch(List<Amenities> amenities) {
+            if(amenities == null || amenities.isEmpty()) {
+                return null;
+            }
+
+            BooleanExpression result = null;
+
+            for(Amenities amenity : amenities) {
+                try {
+                    BooleanExpression currentCondition = store.amenities.contains(amenity);
+
+                    if (result == null) {
+                        result = currentCondition;
+                    } else {
+                        result = result.and(currentCondition);
+                    }
+                } catch (IllegalArgumentException e) {
+                    throw new IllegalArgumentException("Invalid amenity: " + amenity);
+                }
+            }
+
+            return result;
+        }
+    }
