+++ services/inventory-service/src/test/java/com/example/inventory/HealthCheckControllerTest.java
@@ -1,6 +1,5 @@
 package com.example.inventory;
 
-import com.example.inventory.controller.HealthCheckController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
